requirements.txt
openai
langchain
streamlit
langchain_openai
langchain-community
langchain-google-genai
langchain-google-vertexai
llama-index-llms-gemini
llama-index
ollama
langsmith
python-dotenv
youtube-transcript-api
google-api-python-client
faiss-cpu
sentence-transformers